+++ Inventory/src/main/java/com/bartolay/inventory/controllers/CompanyRestController.java
@@ -41,7 +41,7 @@ class CompanyRestController {
 	}
 
 	@RequestMapping(value="/api/companies/{id}", method=RequestMethod.GET)
-	public String getById(@PathVariable Long id) {
+	public String getById(@PathVariable Integer id) {
 		try {
 			Company company = companyRepository.apiFindById(id);
 			return stringUtils.encode(company);
@@ -92,7 +92,7 @@ class CompanyRestController {
 	}
 	
 	@RequestMapping(value="/api/companies/{id}", method=RequestMethod.DELETE)
-	public String delete(@PathVariable Long id) throws RestApiException, JsonProcessingException, UnsupportedEncodingException {
+	public String delete(@PathVariable Integer id) throws RestApiException, JsonProcessingException, UnsupportedEncodingException {
 
 		ApiResponse response = null;
 		try {
